+++ handler_extension_test.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 // Author: FishGoddess
-// Email: fishinlove@163.com
+// Email: fishgoddess@qq.com
 // Created at 2020/03/03 22:48:34
 
 package logit
@@ -25,7 +25,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/FishGoddess/logit/writer"
+	"github.com/FishGoddess/logit/files"
 )
 
 // 测试创建文件日志处理器
@@ -48,7 +48,7 @@ func TestNewFileHandler(t *testing.T) {
 
 // 测试创建随时间间隔滚动的文件日志处理器
 func TestNewDurationRollingHandler(t *testing.T) {
-	logger := NewLogger(DebugLevel, NewDurationRollingHandler(time.Second, os.TempDir(), TextEncoder(), ""))
+	logger := NewLogger(DebugLevel, NewDurationRollingHandler(os.TempDir(), time.Second, TextEncoder(), ""))
 	for i := 0; i < 5; i++ {
 		logger.Info("1. info!!!!!!!! " + strconv.FormatInt(time.Now().Unix(), 10))
 		time.Sleep(time.Second)
@@ -58,7 +58,7 @@ func TestNewDurationRollingHandler(t *testing.T) {
 
 // 测试按照文件大小自动划分日志文件的日志处理器
 func TestNewSizeRollingHandler(t *testing.T) {
-	logger := NewLogger(DebugLevel, NewSizeRollingHandler(64*writer.KB, os.TempDir(), TextEncoder(), ""))
+	logger := NewLogger(DebugLevel, NewSizeRollingHandler(os.TempDir(), 64*files.KB, TextEncoder(), ""))
 	for i := 0; i < 2000; i++ {
 		logger.Debug("debug...")
 		logger.Info("info...")